split crypto and network code to different es6 modules
