+++ Gruntfile.js
@@ -33,7 +33,7 @@ module.exports = function(grunt) {
         es6transpiler: {
             dist: {
                 files: {
-                    'assets/manager-es5.js': 'assets/manager.js'
+                    'assets/manager-es5.js': 'assets/manager-loaded.js',
                 }
             }
         },
@@ -81,6 +81,29 @@ module.exports = function(grunt) {
         }
     });
 
+    grunt.registerTask('load-js-modules', 'Bundles the es6 modules', function() {
+        var transpiler = require('es6-module-transpiler');
+        var Container = transpiler.Container;
+        var FileResolver = transpiler.FileResolver;
+        var BundleFormatter = transpiler.formatters.bundle;
+
+        var container = new Container({
+            resolvers: [new FileResolver(['assets/'])],
+            formatter: new BundleFormatter()
+        });
+
+        container.getModule('manager');
+        container.write('assets/manager-loaded.js');
+    });
+
+    grunt.registerTask('force-transpile', 'Transpile es6 while ignoring warnings', function () {
+        var tasks = ['es6transpiler'];
+
+        // Use the force option for all tasks declared in the previous line
+        grunt.option('force', true);
+        grunt.task.run(tasks);
+    });
+
     grunt.loadNpmTasks('grunt-uncss');
     grunt.loadNpmTasks('grunt-inline-alt');
     grunt.loadNpmTasks('grunt-processhtml');
@@ -91,6 +114,7 @@ module.exports = function(grunt) {
 
     // Default task(s).
     grunt.registerTask('default', [
+        'load-js-modules',
         'es6transpiler',
         'uglify',
         'uncss',
