+++ package.json
@@ -9,6 +9,7 @@
     "grunt": "~0.4.1",
     "grunt-contrib-htmlmin": "^0.4.0",
     "grunt-contrib-uglify": "^0.9.1",
+    "es6-module-transpiler": "~0.10.0",
     "grunt-es6-transpiler": "^1.0.2",
     "grunt-inline-alt": "^0.3.4",
     "grunt-manifest": "^0.4.0",
